get rid of unnecessary bvtTestSetup
